marking contact as read
