+++ src/containers/Chat/ChatMessages/ChatMessages.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useState, useEffect } from 'react';
-import { useQuery, useMutation, useLazyQuery } from '@apollo/client';
+import { useQuery, useMutation, useLazyQuery, useApolloClient } from '@apollo/client';
 import { CircularProgress, Container } from '@mui/material';
 import dayjs from 'dayjs';
 import { Navigate, useLocation } from 'react-router-dom';
@@ -22,6 +22,7 @@ import {
   GROUP_QUERY_VARIABLES,
   GROUP_COLLECTION_SEARCH_QUERY_VARIABLES,
   getVariables,
+  updateContactCache,
 } from '../../../common/constants';
 import { SEARCH_QUERY } from '../../../graphql/queries/Search';
 import { GROUP_SEARCH_QUERY } from 'graphql/queries/WaGroups';
@@ -32,6 +33,7 @@ import {
 import {
   CREATE_AND_SEND_MESSAGE_MUTATION,
   CREATE_AND_SEND_MESSAGE_TO_COLLECTION_MUTATION,
+  MARK_AS_READ,
 } from '../../../graphql/mutations/Chat';
 import { getCachedConverations, updateConversationsCache } from '../../../services/ChatService';
 import { addLogs, getDisplayName, isSimulator } from '../../../common/utils';
@@ -53,6 +55,7 @@ export interface ChatMessagesProps {
 export const ChatMessages = ({ entityId, collectionId, phoneId }: ChatMessagesProps) => {
   const urlString = new URL(window.location.href);
   const location = useLocation();
+  const client = useApolloClient();
 
   const groups: boolean = location.pathname.includes('group');
   const chatType = groups ? 'waGroup' : 'contact';
@@ -80,6 +83,11 @@ export const ChatMessages = ({ entityId, collectionId, phoneId }: ChatMessagesPr
   const [collectionVariables, setCollectionVariables] = useState<any>({});
   const { t } = useTranslation();
   let dialogBox;
+
+  let searchQuery = groups ? GROUP_SEARCH_QUERY : SEARCH_QUERY;
+
+  // get the conversations stored from the cache
+  let queryVariables = groups ? GROUP_QUERY_VARIABLES : SEARCH_QUERY_VARIABLES;
 
   useEffect(() => {
     setShowLoadMore(true);
@@ -139,10 +147,21 @@ export const ChatMessages = ({ entityId, collectionId, phoneId }: ChatMessagesPr
     },
   });
 
-  let searchQuery = groups ? GROUP_SEARCH_QUERY : SEARCH_QUERY;
+  const [markAsRead] = useMutation(MARK_AS_READ, {
+    onCompleted: (data) => {
+      if (data.markContactMessagesAsRead) {
+        updateContactCache(client, entityId);
+      }
+    },
+  });
 
-  // get the conversations stored from the cache
-  let queryVariables = groups ? GROUP_QUERY_VARIABLES : SEARCH_QUERY_VARIABLES;
+  useEffect(() => {
+    if (!groups && entityId) {
+      markAsRead({
+        variables: { contactId: entityId.toString() },
+      });
+    }
+  }, []);
 
   if (collectionId) {
     queryVariables = groups
